requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.2
cssselect2==0.7.0
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.9
django-allauth==0.61.0
django-bootstrap4==24.1
django-classy-tags==4.1.0
django-cms==4.1.0
django-filer==3.1.1
django-filter==23.5
django-formtools==2.5.1
django-oauth-toolkit==2.3.0
django-polymorphic==3.1.0
django-rest-auth==0.9.5
django-sekizai==4.1.0
django-summernote==0.8.20.0
django-taggit==5.0.1
django-tof==0.3.2
django-treebeard==4.7.1
djangocms-admin-style==3.3.0
djangocms-attributes-field==3.0.0
djangocms-bootstrap4==3.0.0
djangocms-icon==2.1.0
djangocms-link==3.1.1
djangocms-picture==4.1.1
djangocms-text-ckeditor==5.1.5
djangorestframework==3.14.0
easy-thumbnails==2.8.5
gunicorn==21.2.0
html5lib==1.1
idna==3.6
importlib-metadata==7.0.1
jwcrypto==1.5.3
kombu==5.3.5
lxml==5.1.0
Markdown==3.5.2
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
redis==5.0.1
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.6.0
zipp==3.17.0
